refactor: rename to AnalyticsService.ts and move into src/public/services
