+++ src/public/services/AnalyticsService.ts
